update hook examples
