--- client/src/services/Queries.ts
@@ -1,7 +1,0 @@
-import { useQuery } from "@tanstack/react-query"
-import UserService from "./UserService"
-
-export const useUsers = useQuery({
-  queryKey: ["allUsers"],
-  queryFn: () => UserService.getUsers()
-})
