+++ client/src/services/User/UserMutations.ts
@@ -1,0 +1,8 @@
+import { useMutation } from "@tanstack/react-query"
+import UserService, { EditUsersBody } from "./UserService"
+
+export function useEditUserMutation(users: EditUsersBody) {
+  return useMutation({
+    mutationFn: () => UserService.editUsers(users)
+  })
+}
