--- client/src/services/UserService.ts
@@ -1,10 +1,0 @@
-import fetchClient from "./OpenApiFetchClient"
-
-const UserService = {
-  getUsers: async function () {
-    const { data } = await fetchClient.GET("/users", {})
-    return data
-  }
-} as const
-
-export default UserService
